+++ odoo/local-src/specific_stock/models/stock_pack_operation.py
@@ -9,7 +9,10 @@ class StockPackOperation(models.Model):
         'Location name', related='location_id.name', store=True, readonly=True
     )
     operator_id = fields.Many2one(
-        'res.users', related='picking_id.operator_id', readonly=True
+        'res.users',
+        related='picking_id.operator_id',
+        readonly=True,
+        store=True,
     )
     location_dest_name = fields.Char(
         'Location dest name',