+++ odoo/local-src/specific_zetes/__manifest__.py
@@ -37,6 +37,7 @@
         'views/res_users.xml',
         'views/res_partner.xml',
         'views/zetes_logger.xml',
+        'views/stock_pack_operation.xml',
         'views/stock_picking_type.xml',
         'views/stock_picking.xml',